modelo_NLP/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu     # obligatorio

flask==2.3.3
spacy==3.7.2
sentence-transformers==2.6.1
pysentimiento==0.7.3
chromadb
python-dotenv
psycopg2-binary==2.9.9     # cliente PostgreSQL
supabase-py==2.4.1     